Some logging clenup.

+++ test/rest/client/src/main/java/unk/test/rest/client/TestRunner.java
@@ -9,7 +9,7 @@ import java.util.Date;
 import java.util.HashMap;
 
 public class TestRunner {
-    private static final Logger log = LoggerFactory.getLogger(TestRunner.class);
+    private final Logger LOGGER = LoggerFactory.getLogger(this.getClass());
     private static final String itemUrl = "http://localhost:8080/t1/item";
 
     private void iterate(RestTemplate restTemplate, int iter) {
@@ -19,26 +19,25 @@ public class TestRunner {
         HashMap<String, String> uriVars = new HashMap<>();
         uriVars.put("path", wrapObj.getPath());
         uriVars.put("dsc", wrapObj.getDsc());
-        log.debug("===> Iteration started:{}", new Date());
+        LOGGER.debug("===> Iteration started:{}", new Date());
         restTemplate.put(itemUrl + "?path=" + wrapObj.getPath()+"&dsc="+wrapObj.getDsc(), null, uriVars);
         wrapObj = restTemplate.getForObject(itemUrl, JSONWrapper.class);
-        if (wrapObj.isIdValid()) {
-            log.debug("{}=> inserted: {}", wrapObj.getId(),wrapObj);
+        if ((wrapObj!=null) && wrapObj.isIdValid()) {
+            LOGGER.debug("{}=> inserted: {}", wrapObj.getId(),wrapObj);
             String newDesc = "DSC_amended";
             restTemplate.put(itemUrl + "?id="+wrapObj.getId()+"&path=" + wrapObj.getPath()+"&dsc="+newDesc,null);
             wrapObj = restTemplate.getForObject(itemUrl, JSONWrapper.class);
-            if (newDesc.equalsIgnoreCase(wrapObj.getDsc())) {
-                log.debug("{}=> updated: {}", wrapObj.getId(),wrapObj);
+            if ((wrapObj!=null) && newDesc.equalsIgnoreCase(wrapObj.getDsc())) {
+                LOGGER.debug("{}=> updated: {}", wrapObj.getId(),wrapObj);
             }
-
             restTemplate.delete(itemUrl + "?id=" + wrapObj.getId());
             JSONWrapper delObj;
             delObj = restTemplate.getForObject(itemUrl, JSONWrapper.class, wrapObj);
             if ((delObj == null) || !delObj.isIdValid()) {
-                log.debug(wrapObj.getId() + "=> deleted");
+                LOGGER.debug(wrapObj.getId() + "=> deleted");
             }
         }
-        log.debug("===> Iteration finished:" + new Date());
+        LOGGER.debug("===> Iteration finished:" + new Date());
     }
 
     void execute(RestTemplate restTemplate) {
@@ -52,7 +51,7 @@ public class TestRunner {
         } while ((wrapObj != null) && wrapObj.isIdValid());
         int maxI=100;
         int maxj=10;
-        log.info("=== Starting performance test for REST client/server ==");
+        LOGGER.info("=== Starting performance test for REST client/server ==");
         for (int j = 0; j < maxj; j++) {
             Date start=new Date();
             for (int i = 0; i < maxI; i++) {
@@ -60,8 +59,8 @@ public class TestRunner {
             }
             Date stop=new Date();
             long tDiff=stop.getTime()-start.getTime();
-            log.info("Time spent for iteration {} (100 records inserted/updated/deleted) is {} Seconds", j, (stop.getTime()-start.getTime())/1000 );
+            LOGGER.info("Time spent for iteration {} (100 records inserted/updated/deleted) is {} Seconds", j, (stop.getTime()-start.getTime())/1000 );
         }
-        log.info("=== Performance test for REST client/server finished ==");
+        LOGGER.info("=== Performance test for REST client/server finished ==");
     }
 }
